Add files via upload
This is the sentiment analysis of NHS tweets, pre and post-Covid 19. Before COVID-19 refers to one year before COVID-19, which is 2019, and one year after COVID-19 refers to a year after the total lockdown was lifted in the UK. This analysis was carried out in December 2022.  The data mining tool used to scrape these tweets from twitter is a third-party tool called Octoparse.